Fix UBO layout, integrate AABB support, fix most compiler warnings, etc

+++ include/J3D/Material/J3DMaterialFactoryV3.hpp
@@ -60,7 +60,7 @@ class J3DMaterialFactoryV3 {
 		T newComp;
 		
 		ptrdiff_t currentOffset = stream->tell();
-		uint32_t pos = offset + (index * newComp.GetElementSize());
+		uint32_t pos = (uint32_t)(offset + (index * newComp.GetElementSize()));
 		stream->seek(pos);
 
 		newComp.Deserialize(stream);
